requirements.txt
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
alembic==1.13.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
av==12.2.0
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-crc32c==1.5.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==9.0.0
pycparser==2.22
pyee==11.1.0
pylibsrtp==0.10.0
pymongo==4.8.0
PyMySQL==1.1.1
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
